pyproject.toml
[tool.poetry]
name = "chatgpt-dall-e"
version = "0.1.0"
description = ""
authors = ["Shkaf4i3 <fleximes@protonmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.13.0"
python-dotenv = "^1.0.1"
openai = "^1.44.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
